requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-ckeditor==5.9.0
django-ckeditor-filebrowser-filer==0.3.0
django-filer==1.7.1
django-jet2==1.0.12
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
djangorestframework==3.11.0
easy-thumbnails==2.7
google-api-python-client==1.4.1
httplib2==0.18.1
idna==2.9
oauth2client==4.1.3
p==0.0.0
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
